+++ docs/about/whats_new.rst
@@ -19,6 +19,8 @@ v1.9.next
 * Fix broken documentation build (:pull:`1668`)
 * Rename `DatasetType` to `Product` in all the tests (:pull:`1671`)
 * Documentation updates for 1.9 release (:pull:`1664`, :pull:`1699`)
+* Fix metadata issues with new Lineage API. (:pull:`1677`)
+* Fix metadata issues with new Lineage API. (:pull:`1679`)
 * Suppress annoying config warning when configuring by environment variable (:pull:`1680`)
 
 v1.9.0-rc11 (28th October 2024)